+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # HDMF Changelog
 
+## HDMF 2.3.0 (Upcoming)
+
+### New features
+- Add ability to specify a custom class for new columns to a `DynamicTable` that are not `VectorData`,
+  `DynamicTableRegion`, or `VocabData` using `DynamicTable.__columns__` or `DynamicTable.add_column(...)`. @rly (#436)  
+- Add capability to add a row to a column after IO. @bendichter (#426)
+
+### Bug fixes
+- Fix handling of empty lists against a spec with text/bytes dtype. @rly (#434)
+- Fix handling of 1-element datasets with compound dtype against a scalar spec with text/bytes dtype. @rly (#438)
+
 ## HDMF 2.2.0 (August 14, 2020)
 
 ### New features